pyproject.toml
[tool.poetry]
name = "present"
version = "0.6.3"
description = "A terminal-based presentation tool with colors and effects."
license = "Apache-2.0"
authors = ["Vinayak Mehta <vmehta94@gmail.com>"]
maintainers = ["Matt White <code@typenil.com>"]
readme = "README.md"
repository = "https://github.com/typenil/present"
classifiers = [
	"Environment :: Console",
	"License :: OSI Approved :: Apache Software License",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3 :: Only",
]

[tool.poetry.scripts]
present = 'present.cli:cli'

[tool.poetry.dependencies]
python = "^3.9"
asciimatics = "^1.12.0"
click = "^7.1.2"
mistune = "2.0.0rc1"
pyfiglet = "^0.8.post1"
PyYAML = "^5.4.1"
loguru = "^0.5.3"
Pygments = "^2.8.0"
Cython = "^0.29.22"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
mypy = "^0.812"
flake8 = "^3.8.4"
pylint = "^2.7.1"
bandit = "^1.7.0"
ipdb = "^0.13.4"

[tool.black]
line-length = 79
target-version = ['py36','py37','py38']
skip-numeric-underscore-normalization = true

[tool.pylint.messages_control]
disable = """
missing-docstring,bad-continuation,fixme
"""

[tool.pylint.reports]
output-format = "colorized"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
